requirements.txt
langchain-mcp-adapters
langchain
langgraph
langchain_community
langchain_openai
openai
httpx

langchain-google-genai

python-dotenv